04/11/2023 - 복습

+++ DataBase/복습(04.11까지)/CRUD/articles/views.py
@@ -57,15 +57,33 @@ def update(request, pk):
     return render(request, 'articles/update.html', context)
 
 def comments_create(request, pk):
-    article=Article.objects.get(pk=pk)
-    commentForm=CommentForm(request.POST)
-    if commentForm.is_valid():
-        comment=commentForm.save(commit=False)
-        comment.article=article
-        comment.save()
-    return redirect('articles:detail', article.pk)
+    if request.user.is_authenticated:
+        article=Article.objects.get(pk=pk)
+        commentForm=CommentForm(request.POST)
+        if commentForm.is_valid():
+            comment=commentForm.save(commit=False)
+            comment.article=article
+            comment.user=request.user
+            comment.save()
+        return redirect('articles:detail', article.pk)
+    return redirect('accounts:login')
 
 def comments_delete(request, article_pk, comment_pk):
-    comment=Comment.objects.get(pk=comment_pk)
-    comment.delete()
-    return redirect('articles:detail', article_pk)
+    if request.user.is_authenticated:     
+        comment=Comment.objects.get(pk=comment_pk)
+        if comment.user==request.user:
+            comment.delete()
+        return redirect('articles:detail', article_pk)
+    return redirect('accounts:login')
+
+def likes(request, article_pk):
+    if request.user.is_authenticated:
+        article=Article.objects.get(pk=article_pk)
+        # Like가 되어있으면 취소
+        if article.like_users.filter(pk=request.user.pk).exists():
+            article.like_users.remove(request.user)
+        # Like가 되어있지 않으면 표시해줌
+        else:
+            article.like_users.add(request.user)
+        return redirect('articles:index')
+    return redirect('accounts:login')
